rename api_server to http_server
