+++ rust/sglang-server/src/http_server/openai/reasoning.rs
