+++ rust/sglang-server/src/http_server/openai/template.rs
