+++ rust/sglang-server/src/http_server/openai/tools.rs
